fix: split team balance client-safe types
