+++ apps/webapp/src/app/[locale]/(app)/team/team-time-balance-types.ts
@@ -1,0 +1,8 @@
+export type EmployeeTimeBalancePayload = {
+	year: number;
+	actualMinutes: number;
+	expectedMinutes: number;
+	absenceAdjustedMinutes: number;
+	balanceMinutes: number;
+	calculatedAt: Date;
+};
